agent: Allow model to provide stringified `timezone` in `now` tool
